backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary # PostgreSQL adapter
daphne>=4.0.0 # ASGI Server (needed for Channels)
channels>=4.0.0
channels-redis>=4.1.0
gunicorn # WSGI server for production (Good practice to include)
python-dotenv # For loading settings from the .env file
djangorestframework-simplejwt
cryptography